feat(baseline): reworking bot commands
