+++ localstack/init.py
@@ -1,9 +1,7 @@
 import boto3
 import os
 
-print("Starting S3 initialization...")
 bucket_name = os.getenv("S3_BUCKET")
-print(f"Creating bucket: {bucket_name}")
 
 s3_client = boto3.client(
     "s3",
@@ -13,4 +11,3 @@ s3_client = boto3.client(
 )
 
 s3_client.create_bucket(Bucket=bucket_name)
-print(f"Bucket {bucket_name} created successfully")
