+++ pkg/clients/clients.go
@@ -1,4 +1,4 @@
-package checks
+package clients
 
 import "strings"
