+++ pkg/discord/config.go
@@ -1,0 +1,6 @@
+package discord
+
+type Config struct {
+	GrafanaToken string
+	DiscordToken string
+}
